pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-lens"
version = "0.1.0"
description = "A description of your project"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "hypha_rpc",
    "hypha",
    "numpy",
    "pillow",
    "httpx",
    "python-dotenv",
    "fastapi",
    "aiobotocore",
]

[project.urls]
homepage = "https://github.com/aicell-lab/agent-lens"
repository = "https://github.com/aicell-lab/agent-lens"

[tool.setuptools.packages.find]
include = ["agent_lens*"]

[tool.setuptools.package-data]
"agent_lens" = ["*.txt", "*.md"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --tb=short"
testpaths = ["agent_lens/tests", "tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "hardware: marks tests that require hardware access",
]